fix(gemini): bound thinking on /complete so Gemini 3 emits JSON (fixes #316)
gemini-3.5-flash (Gemini 3) on /complete had NO thinkingConfig, so it
defaulted to high thinking and burned the entire 64k output budget
reasoning internally — emitting only 7-9 visible tokens before
finishReason: MAX_TOKENS. Every jsonMode /complete 502'd, taking down
brand-service ICP-suggest, human-service audiences-suggest and
extract-fields (onboarding generation dead).

Add `thinkingConfig: buildThinkingConfig(model)` to the /complete body,
the same generation-specific helper the /chat path already uses
(commit 1991af2): thinkingLevel "low" for Gemini 3.x, thinkingBudget
8192 for Gemini 2.5. PR #133's old `thinkingBudget: 0` is NOT a valid
Gemini-3 disable (only accepted for backwards-compat → same degenerate
output); the knob for Gemini 3 is thinkingLevel.

Move buildThinkingConfig from gemini-chat.ts into gemini.ts (single
source of truth, no circular import — gemini-chat already depends on
gemini). Flip the stale "does not send thinkingConfig" test (that
omission WAS the bug) and add Gemini-3-vs-2.5 regression coverage.
