+++ CLAUDE.md
@@ -109,7 +109,11 @@ The number of grounding searches per request can only be bounded on **Anthropic*
 
 `completeWithGemini` (`src/lib/gemini.ts`) MUST set `generationConfig.maxOutputTokens` (default `GEMINI_MAX_OUTPUT_TOKENS` = 64k, or the caller's `maxTokens` when supplied). This is **not** an artificial limit — Gemini falls back to a *lower* per-model **default** cap when the field is omitted, so omitting it gives the model LESS room and truncates long responses sooner. The Anthropic path mirrors it (`anthropic.ts` `MAX_TOKENS = 64_000`, lowered only by an explicit caller `maxTokens` — Anthropic's API *requires* `max_tokens`, which is why only the Gemini side ever silently lost its budget). **Never remove `maxOutputTokens` from the Gemini generationConfig "because the provider manages its own limits"** — that premise is backwards and is exactly what PR #140 did, regressing into truncated-JSON 502s (incident 2026-06-04). The pre-#140 code set it; it was correct. **This provider output CAP is a different number from the cost-provision HOLD — see the next section; do not re-couple them.**
 
-When the model still stops at the budget (`finishReason: "MAX_TOKENS"`), the partial output in **JSON mode** is truncated JSON. `gemini.ts` **throws** `[gemini] Output truncated (MAX_TOKENS)` so the failure surfaces with a clear cause — NOT the cryptic `JSON.parse: Unterminated string` 502 that the downstream strict parse would otherwise emit. The throw decision keys on `jsonMode` (`responseFormat === "json" || responseSchema != null`), threaded into `callGeminiOnce` — do NOT narrow it to `responseFormat === "json"` alone, or `responseSchema`-only callers regress to the cryptic parse error. **Text mode** still returns partial content with a warning (preserves PR #132). This is fail-loud, not jsonrepair — it is consistent with the `/complete` strict-parse contract, not a violation of it. The next lever if huge-output JSON still truncates at 64k is an explicit `thinkingConfig` (Gemini 3 thinking tokens count against the budget) — a deliberate follow-up, not a silent re-removal of the cap.
+When the model still stops at the budget (`finishReason: "MAX_TOKENS"`), the partial output in **JSON mode** is truncated JSON. `gemini.ts` **throws** `[gemini] Output truncated (MAX_TOKENS)` so the failure surfaces with a clear cause — NOT the cryptic `JSON.parse: Unterminated string` 502 that the downstream strict parse would otherwise emit. The throw decision keys on `jsonMode` (`responseFormat === "json" || responseSchema != null`), threaded into `callGeminiOnce` — do NOT narrow it to `responseFormat === "json"` alone, or `responseSchema`-only callers regress to the cryptic parse error. **Text mode** still returns partial content with a warning (preserves PR #132). This is fail-loud, not jsonrepair — it is consistent with the `/complete` strict-parse contract, not a violation of it.
+
+### Gemini THINKING must be bounded on `/complete` — `thinkingConfig: buildThinkingConfig(model)` is REQUIRED (Gemini 3 burns the whole budget thinking)
+
+`completeWithGemini` MUST set `generationConfig.thinkingConfig = buildThinkingConfig(model)` (`src/lib/gemini.ts`). Gemini 3 thinking tokens count against `maxOutputTokens`, and a Gemini 3 model with NO thinkingConfig defaults to HIGH thinking — it spends the entire 64k output budget reasoning internally and emits only 7-9 visible tokens before `finishReason: MAX_TOKENS`, so EVERY jsonMode `/complete` 502s (incident 2026-06-24, #316: brand-service ICP-suggest / human-service audiences-suggest / extract-fields all down → onboarding generation dead). `buildThinkingConfig` is generation-specific: **Gemini 3.x (`gemini-3*`, incl `gemini-3.5-flash`) → `{ thinkingLevel: "low" }`; Gemini 2.5 → `{ thinkingBudget: 8192 }`.** It is the SAME helper the `/chat` path uses (`gemini-chat.ts`, commit 1991af2) — defined once in `gemini.ts`, imported by `gemini-chat.ts` (no circular import: gemini-chat already depends on gemini). **Never remove `thinkingConfig` from the `/complete` body** (that omission is the #316 bug), and **never "disable" Gemini-3 thinking with `thinkingBudget: 0`** (PR #133's value) — on Gemini 3 `thinkingBudget` is only accepted for backwards-compat and produces the same degenerate empty/truncated output; the disable knob for Gemini 3 is `thinkingLevel`, not `thinkingBudget`. This is a SEPARATE lever from the `maxOutputTokens` cap above — raising the cap alone never fixed it (thinking just consumed the larger budget too, exactly why #131's 16k→64k bump didn't help).
 
 ### Cost provision HOLD is right-sized — decoupled from the provider output cap; never revert to a flat 64k
